start parsing messages
I was trying to detect log messages indicating that a heartbeat request or reply was sent or received, so I could correlate them. This doesn't yet account for mongobridge (it fails with an assertion, since the JS test prefixes with e.g. "d20030" a message from a host that thinks it's on port 20020, due to mongobridge). In any case it's the wrong approach, Shiviz doesn't need send/receive events, it just needs to know when a vector clock changes. I'm going to implement a vector clock for MongoDB messages and resume.